@@ -1,28 +1,52 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
+using DataGridWithWebApiService.Data;
+using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.Web;
 
-namespace DataGridWithWebApiService
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
-    }
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+builder.Services.AddRazorPages();
+builder.Services.AddServerSideBlazor();
+builder.Services.AddDevExpressBlazor();
+builder.Services.AddSingleton<WeatherForecastService>();
+
+// Server Side Blazor doesn't register HttpClient by default
+if (!builder.Services.Any(x => x.ServiceType == typeof(HttpClient))) {
+    // Setup HttpClient for server side in a client side compatible fashion
+    builder.Services.AddScoped<HttpClient>(s => {
+        // Creating the URI helper needs to wait until the JS Runtime is initialized, so defer it.
+        var uriHelper = s.GetRequiredService<NavigationManager>();
+        return new HttpClient {
+            BaseAddress = new Uri(uriHelper.BaseUri)
+        };
+    });
 }
+builder.Services.AddScoped<WebServicePath>();
+
+builder.Services.Configure<DevExpress.Blazor.Configuration.GlobalOptions>(options => {
+    options.BootstrapVersion = DevExpress.Blazor.BootstrapVersion.v5;
+});
+builder.WebHost.UseWebRoot("wwwroot");
+builder.WebHost.UseStaticWebAssets();
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (!app.Environment.IsDevelopment())
+{
+    app.UseExceptionHandler("/Error");
+    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+    app.UseHsts();
+}
+app.UseHttpsRedirection();
+
+app.UseStaticFiles();
+
+app.UseRouting();
+
+
+app.MapBlazorHub();
+app.MapFallbackToPage("/_Host");
+
+app.Run();
