@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace DataGridWithWebApiService.Data {
+﻿namespace DataGridWithWebApiService.Data {
     class Categories {
         public int CategoryId { get; set; }
-        public string CategoryName { get; set; }
-        public string Description { get; set; }
+        public string? CategoryName { get; set; }
+        public string? Description { get; set; }
     }
 }
