@@ -1,5 +1,5 @@
-﻿namespace DataGridWithWebApiService {
-    internal class WebServicePath {
+﻿namespace DataGridWithWebApiService.Data {
+    public class WebServicePath {
         public string Path { get { return "https://localhost:44314/api/"; } }
     }
 }
