@@ -20,40 +20,32 @@ else {
         <DxDataGridColumn Field=@nameof(Categories.Description)></DxDataGridColumn>
     </DxDataGrid>
 }
-@functions {
+@code {
     List<Categories> categories;
-
-    protected override async Task OnInitializedAsync()
-    {
+    protected override async Task OnInitializedAsync() {
         categories = await Http.GetFromJsonAsync<List<Categories>>(wsp.Path + "categories");
     }
-    async void OnRowRemoving(Categories dataItem)
-    {
+    async void OnRowRemoving(Categories dataItem) {
         await Http.DeleteAsync(wsp.Path + "categories/" + dataItem.CategoryId);
         categories = await Http.GetFromJsonAsync<List<Categories>>(wsp.Path + "categories");
         await InvokeAsync(StateHasChanged);
     }
-    async void OnRowUpdating(Categories dataItem, Dictionary<string, object> newValues)
-    {
+    async void OnRowUpdating(Categories dataItem, Dictionary<string, object> newValues) {
         SetNewValues(dataItem, newValues);
         await Http.PutAsJsonAsync(wsp.Path + "categories/" + dataItem.CategoryId, dataItem);
         categories = await Http.GetFromJsonAsync<List<Categories>>(wsp.Path + "categories");
         await InvokeAsync(StateHasChanged);
     }
-    async void OnRowInserting(Dictionary<string, object> newValues)
-    {
+    async void OnRowInserting(Dictionary<string, object> newValues) {
         Categories newCategory = new Categories();
         SetNewValues(newCategory, newValues);
         await Http.PostAsJsonAsync<Categories>(wsp.Path + "categories/", newCategory);
         categories = await Http.GetFromJsonAsync<List<Categories>>(wsp.Path + "categories");
         await InvokeAsync(StateHasChanged);
     }
-    private void SetNewValues(Categories dataItem, Dictionary<string, object> newValues)
-    {
-        foreach (var field in newValues.Keys)
-        {
-            switch (field)
-            {
+    private void SetNewValues(Categories dataItem, Dictionary<string, object> newValues) {
+        foreach (var field in newValues.Keys) {
+            switch (field) {
                 case nameof(Categories.CategoryName):
                     dataItem.CategoryName = (string)newValues[field];
                     break;
@@ -64,3 +56,4 @@ else {
         }
     }
 }
+
