@@ -1,10 +1,12 @@
-@using System.Net.Http
+﻿@using System.Net.Http
 @using Microsoft.AspNetCore.Authorization
 @using Microsoft.AspNetCore.Components.Authorization
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
+@using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using DataGridWithWebApiService
 @using DataGridWithWebApiService.Shared
+
 @using DevExpress.Blazor
